+++ src/main/java/modelsDAO/UserDAO.java
@@ -88,10 +88,10 @@ public class UserDAO {
                 ps.setString(5, "01");
                 ps.setString(6, email);
                 ps.setString(7, pass);
-                // Preguntarnos si los ids la mejor forma de tratarlos sería con Strings y no ints.
                 ps.setInt(8, Integer.parseInt(schoolId));
                 ps.setInt(9, Integer.parseInt(courseId));
-                ps.executeUpdate();
+                int rows = ps.executeUpdate();
+                System.out.println(rows);
             } catch (SQLException e) {
                 e.printStackTrace();
             }
@@ -138,49 +138,78 @@ public class UserDAO {
         return usuario;
     }
 
-    /*Función provisional. Falta ver cómo avanza la bbdd para poder avanzar*/
-    public static List<User> getStudentsFromTeacher (int profesorId){
-        List<User> alumnos = new ArrayList<>();
+    /**
+     *
+     */
+    public static List<User> getStudentsFromTeacherID(int teacherID) {
+        Connection con = null;
+        List<User> users = new ArrayList<>();
+        try {
+            con = new Conector().getMySqlConnection();
+            if (con != null) {
+                PreparedStatement preparedStatement = con.prepareStatement("SELECT * FROM user_obj WHERE id = ?");
+                preparedStatement.setInt(1, teacherID);
+                ResultSet resultSet = preparedStatement.executeQuery();
+                while (resultSet.next()) {
+                    users.add(createUser(con, resultSet));
+                }
+            }
 
-        //Añadir sentencia sql para buscar los alumnos en función del id del profesor pasado como parámentro
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        } catch (ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
 
-        return alumnos;
+        return users;
     }
-    
-    
+
+
     /**
      * Método que recupera todos los profesores existentes en la BBDD,
      * es decir todos los usuarios cuyo tipo de usuario sea "02"
      * @author Óscar
      */
-    
-    public static List<String> getAllNameTeachers(){
-      
-      List<String> profesores = new ArrayList<>();
-      
-      try (Connection conx = new Conector().getMySqlConnection()){
-        Statement sentencia = conx.createStatement();
-         
-        //Sentencia sql para traerme todos los profesores
-        String sql = "SELECT user_name, user_surname FROM user_obj WHERE user_type = '02';";
-       
-        ResultSet rs = sentencia.executeQuery(sql);
-        
-        //Se añaden los resultados a la lista de profesores
-        while (rs.next()) {
-          profesores.add(rs.getString(1) + " " + rs.getString(2));
+
+    public static List<User> getAllNameTeachers(int course_id, int school_id) {
+        List<User> teachers = new ArrayList<>();
+        Connection con = null;
+
+        try {
+            con = new Conector().getMySqlConnection();
+            PreparedStatement ps = con.prepareStatement("SELECT distinct * FROM user_obj WHERE (user_type = '02' and course_id = ? and school_id = ?);");
+            ps.setInt(1,course_id);
+            ps.setInt(2,school_id);
+            //Sentencia sql para traerme todos los profesores
+            ResultSet rs = ps.executeQuery();
+            //Se añaden los resultados a la lista de profesores
+            while (rs.next()) {
+                User user = new User();
+                user.setId(rs.getInt("id"));
+                user.setName(rs.getString("user_name") + " " + rs.getString("user_surname"));
+                user.setUserType(rs.getString("user_type"));
+                user.setId_school(rs.getInt("school_id"));
+                user.setId_course(rs.getInt("course_id"));
+                teachers.add(user);
+            }
+            ps.close();
+            rs.close();
+
+        } catch (ClassNotFoundException | SQLException e) {
+            e.printStackTrace();
+        } finally {
+            if (con != null) {
+                try {
+                    con.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            }
         }
-        
-      } catch (ClassNotFoundException | SQLException e) {
-        e.printStackTrace();
-      }
-      
-      //Retornamos la lista con los nombres de los profesores
-      return profesores;
+
+        //Retornamos la lista con los nombres de los profesores
+        return teachers;
     }
-    
-    
-    
 
     /**
      * Método que se utiliza para cambiar la contraseña de un usuario existente
